refactor(kanban): move workflow swimlane selectors out of the component layer
selectWorkflowSwimlanes and selectMobileNavigatorWorkflows exported from
swimlane-container.tsx were imported by the domain hook
use-kanban-display-settings.ts, a layering smell that also meant the
hook's unit test transitively loaded the whole swimlane component tree.

Move both selectors (and the shared WorkflowLike type) into
lib/kanban/workflow-swimlanes.ts, joining the other pure kanban
selectors. filterTasks stays in swimlane-container.tsx — it is the
dual-filter seam pinned there and has no cross-layer importer.

Pure move: no rendered output, ordering, or persisted value changes.
